Rename 'hourglass' to 'calc'. Fixes #93.
