+++ .about.yml
@@ -59,7 +59,7 @@ links:
   text: CALC production instance
 - url: https:/calc-dev.apps.cloud.gov/
   text: CALC development instance
-- url: https://trello.com/b/LjXJaVbZ/hourglass
+- url: https://trello.com/b/LjXJaVbZ/calc
   text: Original project sprint board
 - url: https://trello.com/b/qiJYky5c/calc-data-capture
   text: Data Capture project sprint board
