+++ .env.sample
@@ -1,3 +1,3 @@
 DEBUG=yup
-DATABASE_URL=postgres://localhost/hourglass
+DATABASE_URL=postgres://localhost/calc
 DOCKER_EXPOSED_PORT=8000
