+++ calc/templates/admin/search_form.html
