+++ calc/tests/test_configuration.py
