+++ docs/conf.py
@@ -67,7 +67,7 @@ author = '18F'
 # built documents.
 #
 # The short X.Y version.
-VERSION_PY_PATH = os.path.join(BASE_DIR, 'hourglass', 'version.py')
+VERSION_PY_PATH = os.path.join(BASE_DIR, 'calc', 'version.py')
 _globs = {}
 exec(open(VERSION_PY_PATH).read(), _globs)  # nosec
 version = _globs['__version__']
