--- hourglass/setup.sql
@@ -1,2 +1,0 @@
-ALTER USER hourglass CREATEDB;
-GRANT ALL ON DATABASE hourglass TO hourglass;
