+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-DJANGO_SETTINGS_MODULE=hourglass.settings
+DJANGO_SETTINGS_MODULE=calc.settings
 python_classes=*Test Test*
 python_files=tests.py test_*.py
 addopts = --doctest-modules
