+++ slackbot/bot.py
@@ -6,7 +6,7 @@ from django.conf import settings
 from django.contrib.sites.models import Site
 from django.contrib.staticfiles.storage import staticfiles_storage
 
-from hourglass.site_utils import absolutify_url
+from calc.site_utils import absolutify_url
 
 
 # Timeout, in seconds, after we give up on Slack. Since Slack isn't
